# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Crown & Blade barber shop booking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  name: string;
  slug: string;
  description: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 1 */
  durationMinutes: number;
}

export interface Barber {
  id: number;
  name: string;
  specialty: string;
  bio: string;
  imageUrl: string;
}

export interface Availability {
  date: string;
  serviceId: number;
  barberId: number;
  isClosed: boolean;
  times: string[];
}

export interface PromotionInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  code: string;
}

export interface Promotion {
  valid: boolean;
  code: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  discountPercent: number;
  message: string;
}

export interface BookingInput {
  serviceId: number;
  barberId: number;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  date: string;
  /** @pattern ^\d{2}:\d{2}$ */
  time: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  customerName: string;
  /** @maxLength 160 */
  customerEmail: string;
  /**
     * @minLength 7
     * @maxLength 40
     */
  customerPhone: string;
  /**
     * @maxLength 1000
     * @nullable
     */
  notes?: string | null;
  /**
     * @maxLength 40
     * @nullable
     */
  promotionCode?: string | null;
}

export interface BookingConfirmation {
  reference: string;
  service: Service;
  barber: Barber;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  date: string;
  time: string;
  endTime: string;
  durationMinutes: number;
  location: string;
  subtotal: number;
  discount: number;
  total: number;
  /** @nullable */
  promotionCode: string | null;
}

export interface ContactInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /** @maxLength 160 */
  email: string;
  /**
     * @minLength 10
     * @maxLength 3000
     */
  message: string;
}

export interface ContactConfirmation {
  received: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetAvailabilityParams = {
serviceId: number;
barberId: number;
/**
 * Calendar date in YYYY-MM-DD format
 * @pattern ^\d{4}-\d{2}-\d{2}$
 */
date: string;
};