# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Crown & Blade barber shop booking API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: catalogue
    description: Services and barber profiles
  - name: booking
    description: Availability and appointment booking
  - name: contact
    description: Private contact enquiries
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /services:
    get:
      operationId: listServices
      tags: [catalogue]
      summary: List active barber services
      responses:
        "200":
          description: Service catalogue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
  /barbers:
    get:
      operationId: listBarbers
      tags: [catalogue]
      summary: List active barbers
      responses:
        "200":
          description: Barber profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Barber"
  /availability:
    get:
      operationId: getAvailability
      tags: [booking]
      summary: Get available appointment times
      parameters:
        - name: serviceId
          in: query
          required: true
          schema:
            type: integer
        - name: barberId
          in: query
          required: true
          schema:
            type: integer
        - name: date
          in: query
          required: true
          description: Calendar date in YYYY-MM-DD format
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
      responses:
        "200":
          description: Available times for the selected day
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Availability"
        "400":
          description: Invalid or unavailable date selection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /promotions/validate:
    post:
      operationId: validatePromotion
      tags: [booking]
      summary: Validate a promotional code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionInput"
      responses:
        "200":
          description: Promotion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"
        "400":
          description: Invalid promotion code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings:
    post:
      operationId: createBooking
      tags: [booking]
      summary: Reserve an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Confirmed booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingConfirmation"
        "400":
          description: Invalid booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Selected time was just reserved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Save a private contact enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Contact enquiry saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactConfirmation"
        "400":
          description: Invalid contact details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Service:
      type: object
      required: [id, name, slug, description, price, durationMinutes]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: integer
          minimum: 0
        durationMinutes:
          type: integer
          minimum: 1
    Barber:
      type: object
      required: [id, name, specialty, bio, imageUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        specialty:
          type: string
        bio:
          type: string
        imageUrl:
          type: string
          format: uri
    Availability:
      type: object
      required: [date, serviceId, barberId, isClosed, times]
      properties:
        date:
          type: string
        serviceId:
          type: integer
        barberId:
          type: integer
        isClosed:
          type: boolean
        times:
          type: array
          items:
            type: string
    PromotionInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 40
    Promotion:
      type: object
      required: [valid, code, discountPercent, message]
      properties:
        valid:
          type: boolean
        code:
          type: string
        discountPercent:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
    BookingInput:
      type: object
      required:
        - serviceId
        - barberId
        - date
        - time
        - customerName
        - customerEmail
        - customerPhone
      properties:
        serviceId:
          type: integer
        barberId:
          type: integer
        date:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        time:
          type: string
          pattern: '^\d{2}:\d{2}$'
        customerName:
          type: string
          minLength: 2
          maxLength: 120
        customerEmail:
          type: string
          format: email
          maxLength: 160
        customerPhone:
          type: string
          minLength: 7
          maxLength: 40
        notes:
          type: ["string", "null"]
          maxLength: 1000
        promotionCode:
          type: ["string", "null"]
          maxLength: 40
    BookingConfirmation:
      type: object
      required:
        - reference
        - service
        - barber
        - customerName
        - customerEmail
        - customerPhone
        - date
        - time
        - endTime
        - durationMinutes
        - location
        - subtotal
        - discount
        - total
        - promotionCode
      properties:
        reference:
          type: string
        service:
          $ref: "#/components/schemas/Service"
        barber:
          $ref: "#/components/schemas/Barber"
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        date:
          type: string
        time:
          type: string
        endTime:
          type: string
        durationMinutes:
          type: integer
        location:
          type: string
        subtotal:
          type: integer
        discount:
          type: integer
        total:
          type: integer
        promotionCode:
          type: ["string", "null"]
    ContactInput:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
          format: email
          maxLength: 160
        message:
          type: string
          minLength: 10
          maxLength: 3000
    ContactConfirmation:
      type: object
      required: [received, message]
      properties:
        received:
          type: boolean
        message:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string