# lib/api-zod/src/generated/types/availability.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Crown & Blade barber shop booking API
 * OpenAPI spec version: 0.1.0
 */

export interface Availability {
  date: string;
  serviceId: number;
  barberId: number;
  isClosed: boolean;
  times: string[];
}