# lib/api-zod/src/generated/types/barber.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Crown & Blade barber shop booking API
 * OpenAPI spec version: 0.1.0
 */

export interface Barber {
  id: number;
  name: string;
  specialty: string;
  bio: string;
  imageUrl: string;
}