# lib/api-zod/src/generated/types/bookingConfirmation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Crown & Blade barber shop booking API
 * OpenAPI spec version: 0.1.0
 */
import type { Barber } from './barber';
import type { Service } from './service';

export interface BookingConfirmation {
  reference: string;
  service: Service;
  barber: Barber;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  date: string;
  time: string;
  endTime: string;
  durationMinutes: number;
  location: string;
  subtotal: number;
  discount: number;
  total: number;
  /** @nullable */
  promotionCode: string | null;
}