# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Crown & Blade barber shop booking API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  serviceId: number;
  barberId: number;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  date: string;
  /** @pattern ^\d{2}:\d{2}$ */
  time: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  customerName: string;
  /** @maxLength 160 */
  customerEmail: string;
  /**
     * @minLength 7
     * @maxLength 40
     */
  customerPhone: string;
  /**
     * @maxLength 1000
     * @nullable
     */
  notes?: string | null;
  /**
     * @maxLength 40
     * @nullable
     */
  promotionCode?: string | null;
}