# lib/api-zod/src/generated/types/contactConfirmation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Crown & Blade barber shop booking API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactConfirmation {
  received: boolean;
  message: string;
}