# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Crown & Blade barber shop booking API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /** @maxLength 160 */
  email: string;
  /**
     * @minLength 10
     * @maxLength 3000
     */
  message: string;
}