# lib/api-zod/src/generated/types/promotion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Crown & Blade barber shop booking API
 * OpenAPI spec version: 0.1.0
 */

export interface Promotion {
  valid: boolean;
  code: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  discountPercent: number;
  message: string;
}