# lib/api-zod/src/generated/types/promotionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Crown & Blade barber shop booking API
 * OpenAPI spec version: 0.1.0
 */

export interface PromotionInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  code: string;
}