# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Crown & Blade barber shop booking API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  slug: string;
  description: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 1 */
  durationMinutes: number;
}